Optimize code and supplement test cases
